Add mockData.ts and steps.ts
